requirements.txt
asgiref==3.8.1
certifi==2025.1.31
distlib==0.3.9
Django==5.1.6
filelock==3.17.0
packaging==24.2
pipenv==2024.4.1
platformdirs==4.3.6
python-dotenv==1.0.1
setuptools==75.8.0
sqlparse==0.5.3
virtualenv==20.29.2
djangorestframework==3.14.0
Pillow==10.0.0
